add race, class, and background to review
