+++ src/components/CharacterCreation.svelte
@@ -7,9 +7,6 @@
   import AbilityScoreTab from "components/AbilityScoreTab.svelte";
   import EquipmentTab from "components/EquipmentTab.svelte";
   import SpellTab from "components/SpellTab.svelte";
-  import { capitalize } from "utils/utils.js";
-  import { mergeWith, countBy } from "lodash";
-  import { mergeCustomizer } from "utils/utils.js";
 
   export let closeWindow;
   export let getPosTop;
@@ -86,174 +83,8 @@
     },
   };
 
-  function dataReducer(acc, cur) {
-    console.log(cur);
-    return mergeWith(
-      {},
-      acc,
-      data[cur].data,
-      ...Object.values(data[cur].decisionData).flat(),
-      mergeCustomizer
-    );
-  }
-
-  let mergeData = {};
-
-  $: mergeData = Object.keys(data).reduce(dataReducer, {});
-
   let tabs = ["Bio", "Races", "Class", "Abilities", "Background", "Equipment", "Spells", "Review"];
   let currentTab = "Bio";
-
-  async function createCharacter(event) {
-    event.preventDefault();
-
-    const actor = await Actor.create({
-      name: "New Test Actor",
-      type: "character",
-      img: "icons/svg/mystery-man.svg",
-      sort: 12000,
-      data: {},
-      token: {},
-      items: [],
-      flags: {},
-    });
-
-    let actorData = {};
-
-    if (mergeData.abilities) {
-      for (const [ability, value] of Object.entries(mergeData.abilities)) {
-        actorData[`data.abilities.${ability}.value`] = value;
-      }
-    }
-
-    if (mergeData.save_proficiencies) {
-      for (const ability of mergeData.save_proficiencies) {
-        actorData[`data.abilities.${ability}.proficient`] = 1;
-      }
-    }
-
-    if (mergeData.speed) actorData["data.attributes.movement.walk"] = mergeData.speed;
-
-    if (mergeData.size) actorData["data.traits.size"] = mergeData.size;
-
-    if (mergeData.race) actorData["data.details.race"] = mergeData.race;
-
-    if (mergeData.name) actorData["name"] = mergeData.name;
-
-    if (mergeData.profileImage) actorData["img"] = mergeData.profileImage;
-
-    if (mergeData.tokenImage) actorData["token.img"] = mergeData.tokenImage;
-
-    if (mergeData.alignment) actorData["data.details.alignment"] = mergeData.alignment;
-
-    if (mergeData.appearance) actorData["data.details.appearance"] = mergeData.appearance;
-
-    if (mergeData?.token?.dimSight) actorData["token.dimSight"] = mergeData.token.dimSight;
-
-    if (mergeData.languages) actorData["data.traits.languages.value"] = mergeData.languages;
-
-    if (mergeData.currency) actorData["data.currency"] = mergeData.currency;
-
-    if (mergeData.personality) actorData["data.details.trait"] = mergeData.personality.join("\n");
-
-    if (mergeData.ideal) actorData["data.details.ideal"] = mergeData.ideal.join("\n");
-
-    if (mergeData.bond) actorData["data.details.bond"] = mergeData.bond.join("\n");
-
-    if (mergeData.flaw) actorData["data.details.flaw"] = mergeData.flaw.join("\n");
-
-    if (mergeData.weapon_proficiencies) {
-      const weaponCategories = ["mar", "sim"];
-
-      let value = [];
-      let custom = [];
-
-      for (const weaponProf of mergeData.weapon_proficiencies) {
-        if (weaponCategories.indexOf(weaponProf) > -1) {
-          value.push(weaponProf);
-        } else {
-          custom.push(weaponProf);
-        }
-      }
-      actorData["data.traits.weaponProf.value"] = value;
-      actorData["data.traits.weaponProf.custom"] = custom
-        .map((weaponProf) => capitalize(weaponProf))
-        .join(";");
-    }
-
-    if (mergeData.armor_proficiencies) {
-      const armorCategories = ["lgt", "med", "hvy", "shl"];
-
-      let value = [];
-      let custom = [];
-
-      for (const armorProf of mergeData.armor_proficiencies) {
-        if (armorCategories.indexOf(armorProf) > -1) {
-          value.push(armorProf);
-        } else {
-          custom.push(armorProf);
-        }
-      }
-      actorData["data.traits.armorProf.value"] = value;
-      actorData["data.traits.armorProf.custom"] = custom
-        .map((armorProf) => capitalize(armorProf))
-        .join(";");
-    }
-
-    if (mergeData.tool_proficiencies) {
-      const toolCategories = [
-        "art",
-        "disg",
-        "forg",
-        "game",
-        "herb",
-        "music",
-        "navg",
-        "pois",
-        "thief",
-        "vehicle",
-      ];
-
-      let value = [];
-      let custom = [];
-
-      for (const toolProf of mergeData.tool_proficiencies) {
-        if (toolCategories.indexOf(toolProf) > -1) {
-          value.push(toolProf);
-        } else {
-          custom.push(toolProf);
-        }
-      }
-      actorData["data.traits.toolProf.value"] = value;
-      actorData["data.traits.toolProf.custom"] = custom
-        .map((toolProf) => capitalize(toolProf))
-        .join(";");
-    }
-
-    if (mergeData.skill_proficiencies) {
-      for (const skill of mergeData.skill_proficiencies) {
-        actorData[`data.skills.${skill}.value`] = 1;
-      }
-    }
-
-    await actor.update(actorData);
-    if (mergeData.items) {
-      const itemUuids = mergeData.items.concat(mergeData.features);
-      const quantities = countBy(itemUuids);
-      const items = await Promise.all(Object.keys(quantities).map((uuid) => fromUuid(uuid)));
-
-      const itemObjects = [];
-      items.forEach((item) => {
-        const itemObj = item.toObject();
-        itemObj.data.quantity = quantities[item.uuid];
-        itemObjects.push(itemObj);
-      });
-
-      await game.dnd5e.entities.Item5e.createDocuments(itemObjects, { parent: actor });
-    }
-
-    closeWindow();
-  }
 </script>
 
 <div>
@@ -295,7 +126,7 @@
   {/if}
 
   {#if currentTab === "Review"}
-    <ReviewCharacterTab data={mergeData} />
+    <ReviewCharacterTab {data} {closeWindow} />
   {/if}
 </div>
